+++ R/file_formatting.R
@@ -1,14 +1,14 @@
 
 # IntCal 98
 ### The files were made by reading the files of the calibration curves as distributed in http://intcal.org/curves:
-# cc.terr <- read.table( "intcal98.14c", header=FALSE, skip=11)
+# cc.terr <- read.table( "~/Downloads/intcal98.14c", header=FALSE, skip=11)
 # cc.terr[,1] = 1950-cc.terr[,1] # since the calendar ages are in BC/AD!
-# cc.marine <- read.table( "marine98.14c", header=FALSE, skip=11)
+# cc.marine <- read.table( "~/Downloads/marine98.14c", header=FALSE, skip=11)
 # cc.marine[,1] = 1950-cc.marine[,1] # since the calendar ages are in BC/AD!
 
 ### ... and were then written in the desired format (increasing cal BP ages, 3 columns, tab separator):
-# write.table( cc.terr[nrow(cc.terr):1,1:3], file="3Col_intcal98.14C", row.names=FALSE, col.names=FALSE)
-# write.table( cc.marine[nrow(cc.marine):1,1:3], file="3Col_marine98.14C", row.names=FALSE, col.names=FALSE)
+# write.table( cc.terr[nrow(cc.terr):1,c(1,4,5)], file="3Col_intcal98.14C", row.names=FALSE, col.names=FALSE)
+# write.table( cc.marine[nrow(cc.marine):1,c(1,4,5)], file="3Col_marine98.14C", row.names=FALSE, col.names=FALSE)
 
 
 # IntCal04
@@ -53,3 +53,23 @@
 #write.table( cc.terr[nrow(cc.terr):1,1:3], file="3Col_intcal20.14C", row.names=FALSE, col.names=FALSE)
 #write.table( cc.marine[nrow(cc.marine):1,1:3], file="3Col_marine20.14C", row.names=FALSE, col.names=FALSE)
 #write.table( cc.south[nrow(cc.south):1,1:3], file="3Col_shcal20.14C", row.names=FALSE, col.names=FALSE)
+
+
+# postbomb curves:
+
+# NH1, NH2, NH3, SH1-2 and SH3 postbomb curves from Hua et al. were copied from the clam and rbacon packages
+
+# kure <- read.table("inst/extdata/Kure.14c", skip=11, header=FALSE)
+# as.c <- array(pMC.age(kure[,2], kure[,3], ratio=1), dim=c(nrow(kure),2))
+# kure <- cbind(kure[,1], as.c)
+# write.table(kure, "inst/extdata/Kure.14C", row.names=FALSE, col.names=FALSE, sep="\t")
+
+# levin <- read.table("inst/extdata/Levin-Kromer.14c", skip=11, header=FALSE)
+# as.c <- array(pMC.age(levin[,2], levin[,3], ratio=1), dim=c(nrow(levin),2))
+# levin <- cbind(levin[,1], as.c)
+# write.table(levin, "inst/extdata/LevinKromer.14C", row.names=FALSE, col.names=FALSE, sep="\t")
+
+# santos <- read.table("inst/extdata/Santos.14c", skip=11, header=FALSE)
+# as.c <- array(pMC.age(santos[,2], santos[,3], ratio=1), dim=c(nrow(santos),2))
+# santos <- cbind(santos[,1], as.c)
+# write.table(santos, "inst/extdata/Santos.14C", row.names=FALSE, col.names=FALSE, sep="\t")