+++ inst/extdata/try1.R
@@ -1,1 +1,0 @@
-dat <- read.table("inst/extdata/intcal20_data.txt", sep="\t")